fix incorrect plants_lib api call for anthills
The aboid node was not correctly set resulting in metric assloads of
anthills, which was clearly not intended :-)

This also alises the current ones to air and replaces them with a new
node name so as to clean up the overabundance.  The alias and node name
can be inverted later after it's certain that all old nodes have been
cleaned out.